+++ release-notes.md
@@ -2,7 +2,8 @@
 
 ## 0.10.0
 - Extend new TypeConstantValidator base class
-- **Incompatible** Moved [TenantStreamId](api/src/main/java/org/fuin/esc/api/TenantStreamId.java) to API 
+- **Incompatible** Moved [TenantStreamId](api/src/main/java/org/fuin/esc/api/TenantStreamId.java) to API
+- **Incompatible** Changed [ProjectionJavaScriptBuilder](spi/src/main/java/org/fuin/esc/spi/ProjectionJavaScriptBuilder.java) for tenant projections
 
 ## 0.9.0
 - Added new `findAll()` method to [SerializedDataTypeRegistry](api/src/main/java/org/fuin/esc/api/SerializedDataTypeRegistry.java)